refactor(runtime): split orchestrator/tests — extract streaming tests
Extract the 5 streaming-path tests into a new `tests/streaming.rs`,
combining what was the "Streaming tests" and "Thinking event streaming
tests" sections of the original monolithic file:

- `run_turn_with_tools_streaming_emits_tokens_through_sink`
- `run_turn_with_tools_streaming_tokens_arrive_in_order`
- `worker_routes_ext_plus_sink_to_streaming_path`
- `run_turn_streaming_emits_thinking_event`
- `run_turn_streaming_emits_thinking_before_tool_calls`

The two helpers `ollama_streaming_body` and
`ollama_tool_calls_with_thinking` move with the tests since they're
streaming-specific.

cargo test -p assistant-runtime --lib reports 220 passed (same as
before); `orchestrator::tests::streaming::*` subtotal is 5.

Third slice in the stack.
